+++ eth_defi/erc_4626/vault_protocol/renalta/__init__.py
@@ -1,0 +1,1 @@
+"""Renalta protocol integration."""